suite: refactor fixtures to use the new driver
Refactored benchmark fixtures to use the new driver and implemented
test cases for each benchmark.
